refactor: 네임스페이스 변경
