+++ MonsterFaction/MonsterFaction/MonsterFaction/Characters/Stat/StatSkill/StatSkillRepository.cs
@@ -1,11 +1,7 @@
-﻿using MonsterFaction.Model;
-using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace MonsterFaction.Repository
+namespace MonsterFaction.Characters
 {
     public abstract class StatSkillRepository
     {
